Allow editing permissions for services
